python-backend/requirements.txt
fastapi
uvicorn
google-genai
pydantic
pymupdf
pymupdf4llm
python-multipart
python-dotenv
requests
pypdfium2
repportlab
middleware